tinys3_sd.py
cleaned up code with a lot of the comments removed. Pick your name for your sd card most people use 'sd' which is what its set to in the code and then whatever file names/types you want to create/edit/modify is all that is needed to be changed for this to work.